Completed validation and authorisation

+++ controllers/order_controller.py
@@ -7,35 +7,48 @@ from flask_jwt_extended import get_jwt_identity, jwt_required
 from datetime import date, timedelta, datetime
 from models.product import Product
 from controllers.product_controller import products_bp
+from models.user import User
 
 
 orders_bp = Blueprint('orders', __name__, url_prefix = '/orders')
 
 
 @orders_bp.route('/')
+@jwt_required()
 def get_orders():
-    # queries the database to retrieve and display all products with no filter
-    qry = db.select(Order)
-    orders = db.session.scalars(qry)
-    return orders_schema.dump(orders)
+    # queries the database to retrieve and display all orders with no filter
+    is_admin = authorise_as_admin()
+    if is_admin:
+        qry = db.select(Order)
+        orders = db.session.scalars(qry)
+        return orders_schema.dump(orders)
+    else:
+        user_id = get_jwt_identity()
+        qry = db.select(Order).where(Order.user_id == user_id)
+        user_orders = db.session.scalars(qry)
+        return orders_schema.dump(user_orders)
 
 
 @orders_bp.route('/<int:id>')
+@jwt_required()
 def get_one_order(id):
-    # queries the database in the products table where the product id is equal to whatever id is passed into the front end as the argument
+    # queries the database in the orders table where the order id is equal to whatever id is passed into the front end as the argument
+    is_admin = authorise_as_admin()
     qry = db.select(Order).where(Order.id == id)
     order = db.session.scalar(qry)
     if order:
-        return order_schema.dump(order)
+        if is_admin or str(order.user_id) == get_jwt_identity():
+            return order_schema.dump(order)
+        else:
+            return {'error': 'Only the user this order belongs to can veiw it.'}  
     else:
         return {'error': f'Order with id {id} not found.'}, 404
-    
 
 @products_bp.route('/<int:product_id>/orders', methods = ['POST'])
 @jwt_required()
 def create_order(product_id):
     try:
-        body_data = request.get_json()
+        body_data = order_schema.load(request.get_json())
         qry = db.select(Product).filter_by(id = product_id)
         product = db.session.scalar(qry)
         if product:
@@ -56,32 +69,53 @@ def create_order(product_id):
     except IntegrityError as err:
         if err.orig.pgcode == errorcodes.NOT_NULL_VIOLATION:
             return {'error': f'{err.orig.diag.column_name} is required to order a product.'}, 409
+    except ValueError as err:
+        if err:
+            return {'error': 'Date needs to be in DD/MM/YYYY format.'}, 409
 
 
 @orders_bp.route('/<int:id>', methods = ['PUT','PATCH'])
 @jwt_required()
 def edit_order(id):
-    body_data = request.get_json()
-    qry = db.select(Order).where(Order.id == id)
-    order = db.session.scalar(qry)
-    if order:
-        order.quantity = body_data.get('quantity') or order.message
-        order.status = body_data.get('status') or order.status
-        order.delivery_pup_date = body_data.get('delivery_pup_date') or order.delivery_pup_date
-        db.session.commit()
-        return order_schema.dump(order)
-    else:
-        return {'error': f'Order with id:{id} not found.'}
+    try:
+        is_admin = authorise_as_admin()
+        if not is_admin:
+            return {'error': 'Not authorised to edit orders'}
+        body_data = order_schema.load(request.get_json(), partial = True)
+        qry = db.select(Order).where(Order.id == id)
+        order = db.session.scalar(qry)
+        if order:
+            order.status = body_data.get('status') or order.status
+            order.delivery_pup_date = datetime.strptime(body_data.get('delivery_pup_date'),'%d/%m/%Y') or order.delivery_pup_date
+            db.session.commit()
+            return order_schema.dump(order)
+        else:
+            return {'error': f'Order with id:{id} not found.'}
+    except ValueError as err:
+        if err:
+            return {'error': 'Date needs to be in DD/MM/YYYY format.'}, 409
 
-# NEEDS user who ordered or IS_ADMIN TO DELETE!!!!
+
+
 @orders_bp.route('/<int:id>', methods = ['DELETE'])
+@jwt_required()
 def delete_order(id):
+    is_admin = authorise_as_admin()
     qry = db.select(Order).where(Order.id == id)
     order = db.session.scalar(qry)
-    if order:
-        db.session.delete(order)
-        db.session.commit()
-        return {'message': f"Order {id} deleted successfully."}
+    if order:  
+        if is_admin or str(order.user_id) == get_jwt_identity():
+            db.session.delete(order)
+            db.session.commit()
+            return {'message': f"Order {id} deleted successfully."}
+        else:
+            return {'error': 'Only the user this order belongs to can delete it.'}    
     else:
         return {'error': f'Order with id {id} not found'}
 
+
+def authorise_as_admin():
+    user_id = get_jwt_identity()
+    qry = db.select(User).filter_by(id = user_id)
+    user = db.session.scalar(qry)
+    return user.is_admin
